Current progress on adding cmake (DO NOT MERGE!)
This is the start of transitioning to [CMake](http://cmake.org) to allow
for anyone to use the build setup of their choice. It is still not
complete.

Currently most of the Zelda Classic target is set up. The unity files are
not included yet. I'm concerned about how to add source files and yet NOT
compile them. Then again, I don't think the non MSVC builds did that. I'm
holding off.

Also, Allegro assumes that MSVC does not include some header files. Either
Allegro needs updating and/or the definitions must be added before Allegro
normally does so.

Feel free to make suggestions or improvements in comments and forks as
usual.