Rewrote constants.
Constants are now treated as normal (unassignable) variables up through
type checking. If at that point the constant's value is known, it is
stripped out of the AST and its value is saved to the symbol table,
similar to how it worked before.

If the constant's value is not known at compile time, it is instead
treated as a normal, unassignable variable.

Also, since they are "normal" variables, constants can be declared in
any scope, not just global.