New Itemdata ----------------- This is a build with new itemdata vars, that I cannot manage to get the packfile to write, despite using version syntax.
I need some help getting this to work, as it is part of the updated
weapon/item editor.

The editor pane is likely ready to use, but I cannot test it past saving
a quest, as loading a quest saved with the packfile changes causes ZQuest
to crash (bad token reading items).

I have done this stuff before, but I am clearly missing something.