refactor: Make functions own compiled code
ZScript::Function objects now track their own vector<Opcode*> code
instead of having it passed around in a map.

Moved builtin code generation (GlobalSymbols.cpp) earlier in the build
process to when the function objects for that code are first created.

Builtin code is no longer specially added to the result. It is instead
pulled from the program structure along with everything else.