Feat & Fix: Item Editor and Combo Editor Changelog: Fixed a packfile issue resulting from V_COMBOS not being set to 10. Changed how magic cost works. Now, the counter used for this is set in the item editor, and ZC will use that counter. In older quests, the counter is automatically set to RUPEES, if the flag 'Use Ruppes instead of Magic' was checked. I have now removed this flag.
The ZC behaviour for checking magic costs has also changed. If the MAGIC
counter is set, ZC checks for the qr_ENABLEMAGIC QR. If that is not set,
ZC defaults to rupees.

If the counter is set to RUPEES< this is always used.

Likewise, if the counter is anything else, the actual counter is used, as
it was not possible to use any other counter in older versions of ZQuest.

The change is in Link.cpp, to the functions checkmagiccost() and paymagiccost().

The ZScript interface to these is itemdata->CostCounter. Thus, itemdata->MagicCost
is no longer appropriate in most cases. I have added a new variable, as
itemdata->Cost that performs as itemdata->MagicCost, to match the new syntax.

V_COMBOS is now 10.
V_ITEMS is now 34.