Feat: Added Link->HitBy[4] and Link->Defense[256] Changelog: Added a pair of Link class arrays. The first, stores the last weapon or npc that hit Link. This is not yet cleared every frame. The values only change when Link is hit.  The ZScript interface for this is Link->HitBy[4], where index 0 is the last npc screen index that hit him, and index [1] is the last eweapon.
The second, is Link->Defense[256], and while the ZScript array is r/w.
there is no code to swet up defences in the quest, nor anything that makes
use of them.

They are not yet saved at all, and this will need an Init Data panel, plus
code to save it with the quest, and to load it. Heck, it'll need new
section info. :(

It will also need code to modify the damage value of weapons that hit Link.