feat: Add #option trace for disabling tracing calls
A call to "#option trace off" will disable all "tracing" function
calls.

Currently, "tracing" functions are defined as those which return void
and start with "Trace" or "print". Later we should change this out to
a specific designation we can attach to a function.

-- Internals --

ASTStmt nodes can now be disabled, meaning that they won't generate
any opcodes. Currently only used for void function calls, but there's
no guarantee what'll happen if an expression in the middle of a
statement is disabled.