Feat: Expanded Tiles -> Game Icons
Changelog: Finished porting the sizeof(word) to sizeof(int) change for tiles
to all struct elements in zdefs, and the code to ahndle the new sizes.

Some minor bugs/glitches remain, including the EDITOR DIALOGs not jumping
to the correct tile page when the tile set is on a page higher than 251, but
all quest loading and runnin SEEMS to be fine at this time.

This needs testing with more quests, particularly older quests, to find any issues.