Fix: Vires without 'Plit on Hit' Set.
Changelog: Previously, if an enemy with Vire movement, but without the death
type 'Split on Hit', would be stunned perpetually in mid-air when hit.

This corrects that behaviour, so that they act normally.

Vires WITH Split on Hit are unaffected by this change; and I did not touch
the effect that knocks them across the screen, as I do not recall if this is
NES behaviour.

We can look into that, at another time.