Feat: Allow Setting or Clearing Catchall on Roomtype None
Changelog: If the Room Type is set to rNONE, in previous versions, the
Catchall-Option was disabled.

Now., the Catch All option allows setting a manual value to the catchall
value for the mapscreen. This permits either clearing it, or manually setting it.

Either is primarily useful for ZScript integration; e.g., a script that reads catchall
where the room type changes.