Feat & Fix: Added Cheat-only keyfile; fixed keyfile issues.
Changelog: This adds a third level of keyfile: .zcheat.
This uses a pseudo-encrypted password in the ASCII text output, so that
a questmaker may share it with playtesters without revealing the master
quest password for editing in ZQuest.

I might touch up the ZQuest Editing (only) password later, to do the same,
using a different mask.

I also did a minor refactor on keyfile behaviour, with some ugly jump points,
and fixed a crash caused by trying to read a file that doesn't exist (typo).