Feat: Initial Work on Invisible Enemy Behaviour Flags
Changelog: Added flags to all enemies (except Ganon):
Enemy is Completely Invisible
Enemy Is Revealed by Item Specified in dmisc13
---> This eventually needs to be a class when positive, and a specific item when negative.
Draw Invisible as Cloaked
Render Cloaked Instead of VISIBLE

All of the above seem to work.

For GANON:
Visible if the player has Level 2 Amulet or higher.
Specific item dispels invisibility (dmisc13)
When not invisible, show shots.
Draw cloaked when not invisible.

This is only PARTLY working. Ganon's shots render, but Ganon does not.

On the Attributes tab for all enemies, Attributes[12] (misc13) is now
for the item that removes invisibility from that enemy, and is listed accordingly.