GUI Fix: Changed colors coded for jwin_radio_proc, jwin_check_proc, and draw_checkbox
These were hard-coded to use jcDARK for the foreground and jcLIGHT for the background; I changed this to jcTEXTFG and jcTEXTBG respectively
I changed the "shadow" on jwin_radio_proc from jcMEDDARK/jcDARK to a hard-coded vc(0) (Black). This is so that it looks proper in both "Light" and "Dark" themes.