Fix: Increased VBound range for tint sums.
Changelog: Changed VBound from -63,63 to -32768,32767.
This allows values beyond what have a direct effect to be stored, allowing for more precision to be kept on the script side.
The new bound is equivalent to the size of `short int`, and cannot be expanded further. It shouldn't need to be expanded further, ever.