Fix: Mapdata FFC Array Sizes
Changelog: Changed array sizes for mapdata ffc arrays from size 32 to size 33.
Index[0] will now return an error.
This is to ensure uniformity between the old ffc namespace and the mapdata namespace:
FFC uses a range of 1 to 32, and mapdata  (previously using a range of 0 to 31), is now identical.
-Z

Also fixed compiler error in previous commit -V