Fix: Don't Draw Some Enemies
Changelog: I had an open if statement in enemy::draw, in the statement:
`if (dontdraw()`

The statement was: `if (family == eeGANON)`, and this needed its own return instruction
to prevent drawing enemies that should always be invisible.

I also added old_draw(), in the event that we need to reference it in the future.

Further, we need to update enemy::drawblock() for enemies larger than 1x1.