Feat/Experiment: Perpetual item Scripts (Initial)
I decided to try to implement item scripts that run perpetually.

While I figured out how to make them run, I have yet to fingure out how
to properly control the loop so that it releases control back to the player
and so that it only occurs when there are item scripts that are properly
initialised by the player.