Fix: Probably Fix Item Script Stack Corruption
Changelog: When Link.cpp calls a new item script, it also clears the script
stack. This should prevent multiple instances of a script tryint to use the
same stack at different stack frames.

It seems to work now, but obviously, it needs more testing, particularly for
collect scripts, and with the QR off.