Modules: Add Name Entry Graphics to Module Format
Changelog: The quest icons (all TEN supported quests) are now set by the
module, along with their csets and X positions.

The file select frame tile, and cset is now set by the module.

The colour shift on quest 3+ is now in the module format, but disabled,
pending other tests.

Updated the classic module file with the new vars.