NPC Scripts: Move Protected Enemy Functions to Public
Changelog: Moved a number of npc functions from protected to public, so that
I can make npc-> ZScript functions for them, to use with npc scipts.

e.g. to add, npc->CanMove(), or npc->ConstantWalk(), so that an npc script
can call `if ( this->CanMove(this->Dir) ) this->ConstantWalk(); `

It's better IMO to call the engine functions for npcs, then to mimic them in ZASM.