Chore: Add FFCore.Waitframe() and replace stupid advanceframe() duplicates.
Changelog: Added a function to FFCore that acts similarly to advanceframe().
but that can safely be called without regard to the status of the global
var `Quit`.

Removed advanceframe2() and advanceframe3(), and changed the spots
that called these to call FFCore.Waitframe().

This is needed to run the onWin and the onDeath scripts for Link.