Feat: Add Compatibility Menu to ZC Player
Changelog: Added a new 'Compatibility' menu to ZC Player.
At present, it has one entry (1.92b163 Warps), but this does nothing.

I enabled this(to function) in a special build of 2.53, and I'm waiting on an
end-user to report if it fixes the issue with wars in old 1.92 quests
before I enable the live code in 2.55.

Other live compatibility patches will go in this menu.

I also enabled the little tick mark icon to appear in the
menus for `Show Debug Console`.