Feat: Emulation Menu Improvements (2.53)
Changelog: Added a var to which the quest loader stores the version, and the build
of ZQuest in which a quest was last saved so that we can read it at any time.

Changed the 'Compatibility' menu to 'Emulation' to match 2.55.

Renamed entry [0] to 'Flip-Flip Cancel and Wave Warps'

Disabled this option if the current quest was made in 2.10 or newer.