Feat: Add Game->Continue() and bool Game->SkipCredits to ZScript
Changelog: Added Game->Continue() to ZScript. This should function
in the same way as Game->End(), but it should cause the game to
resume as with the Continue option in the normal game over menu.

Added bool Game->SkipCredits to ZScript. This should cause the game
to skip over the WIn Game credits, but still advance the quest, normally.