Array Draws: Copy block from historical/old-master
Changrelog: Fixed one issue with array draws, for Lines()
where the array iterator was wrong.

Copied a code block from historical/old-master where Gleeok first
tried to set this up.

The issue is that the array becomes invalid (scope) prior to
the enqueued draws taking place, so there is no data for them to use.

The values need to be pushed into a vector and used, later.