DIALOGUE EXAMPLE & Fix: Example to Select by Tab in Header Dialogues
Changelog: Fixed the quest header dialogue so that the selected object is
activated when the user presses the enter key.

This is an example of how to update dialogues, so that the selected object
(selected using the TAB key), is activated.

You set the initial focus in zc_popup_dialog(DIALOGUE, int focus_element);
Then, you loop through elements, find the selected one with the flag D_GOTFOCUS.
Once you find it, set ret = obj.