Fix: Palette Cycle Issues
Changelog: Palettes above 256 were not set up to have palette cycles in 2.50.
The array in QMisc is [256][3], where it should have been [512][3].

Because of this, palettes above 0xFF had rubbish data associated to their cycle data.

To remedy this, I have done the following:

All palettes > 255 have a null palette loaded into them, in the editor and as
palette data, to zero out their cycle information.

This should occur both in ZC, and in ZQuest.

I disabled clicking the Cycle button in the Level palette editor, and display
a jwin-alert that the palette doesn't support cycling.

Let's hope this doesn't cause crashes in any quests, in ZC.