Feat: Default and Redirected Weapon Types and Defences
Changelog: Ported the New Weapon Defence System that I wrote for the DD 2.60
old master branch, to 2.55.

This does the following:

Items/lweapons now have a 'UseDefence' property.

Effects: If an enemy in the Enemy Editor has its defence category for that
weapon type set to 'None', and the weapon has a base defence type, then
that defence type (effect) will be applied as if the enemy had it set.

Items/lweapons now have a 'UseType' property.
Effects: When hit with this weapon type, the enemy will use the defence
category of the simulated 'UseType' weapon type, rather than the true weapon
type of the weapon.