Feat: Begin Adding Custom run()
Changelog: Added the ability to set the function label for void run()
in ZQuest, and checked for it in the parser.

If this label is set to run(), it all works; but if it is set to something
else, then there is another check in the parser, probably the error handler,
that I missed.