Rip Out Bombchu Quest Rule
Changelog: Lut was wrong... Bombchu enemies do not need a QR to use a superbomb
explosion. They use the weapon set to them.

In older builds, their behaviour was different, hence needing the emulation
rule, but if the user saves a quest in 2.53.x, they can set the enemy weapon
to select the death explosion weapon.