Changed MaxDraws and NumDraws to function style.
Changelog: Changed Graphics->MaxDraws to Graphics->MaxDraws() and changed
Graphics->NumDraws to Graphics->NumDraws().

The change is purely for the purpose of semantics. If these appear to be a variable,
then the user may try to write to them. In the form of a function, it is immediately
clear that they are read-only.

The actual GlobalSymbols code still sees both as a GETTER, as a function with a return value
and no parameters can be a basic GETTER. The only difference, is whether the table entry
precedes the string with "get" or "set".

i.e.
"getMaxDraws" as a type of GETTER creates a variable 'MaxDraws'.
By using the string "MaxDraws", the parser knows that this is the FORM of a function,
'MaxDraws()', but it needs no params to push/pop, so it can still work as a
GETTER type, instead of as a FUNCTION type.

See 'NumNPCs' for more information.