Fix: Offsets for bitmap drawing commands.
Changelog: All bitmap drawing commands now check if the dest is not the
screen, and set their yoffset to 0 if it is not the screen.

They also check if the dest is ! -2, which is the present RENDER TARGET,
but they do not also validate that the RT is not RT_SCREEN.

We may need to do that, too.