Fix: Bitmap->Blit and Bitmap->Read
Chanhgelog: Fixed the type and internal value for the target of Blit.
Before, it would not allow passing a bitmap pointer to the blit target, and
the value for the pointer would have been wrong, in any case.

This fixes both of these issues.

Fixed ZC Player palette corruption when calling bitmap->Read()