Feat: begin Adding Experimental DrawScreenObj functions to bitmap->
Changelog: Began adding commands to the bitmap-> pointer, similar to DrawScreen,
that when done, should draw blocks to the space on a bitmap that a combo occupies,
that the user can utilise in conjunction with GetPixel() to rapidly read values
for solidity mask, type, flag, and inh flag.

This would make it easier to produce large-scale scrolling environments, including
with Mode-7 style effects, as the end-user can blit out these masks for an entire map
to bitmaps, then read the same absolute position on a set of bitmaps to get
the properties of any specific point/node/coordinates.