Feat: Add the ability to get combo info using bitmaps.
Changelog: Added the following bitmap drawing commands:

DrawScreenSolidity() : Draws a palette index based on the raw literal value
for the solidity of each combo space, by drawing a 16px block to that space
in the palette value equal to the solidity mask of that combo.

DrawScreenComboTypes(), as above, but the palette index is equal to the
combo Type ID value.

DrawScreenComboFlags(), as above, for placed flags.

DrawScreenComboIFlags(), as above, for inh. flags.