Parser: Add Opcodes for internal optimization
CASTBOOLI: Sets the value of a register to 1, unless it is currently 0.
CASTBOOLF: Sets the value of a register to 0.0001, unless it is currently 0.
SETTRUEI: Same as SETTRUE, but uses 1 instead of 0.0001.
SETFALSEI: Same as SETFALSE, but uses 1 instead of 0.0001.
SETMOREI: Same as SETMORE, but uses 1 instead of 0.0001.
SETLESSI: Same as SETLESS, but uses 1 instead of 0.0001.

Using these opcodes, relational operators have been internally optimized.