Feat: Add Epilepsy Protection Option
Changelog: Added an epilepsy protection topggle for Triforce pick-up f;ashing.
This is both a toggle in the Player UI, under the Settings Menu:
(Epilepsy Flash Reduction), and it is available as a global Quest Rule.

If enabled as a QR, then there is no way to disable it. The Settings menu option
only enables this protection; it does not flip-flop the QR bit.

This is intentional, as to prevent a user with this setting enabled playing a quest with
the rule enabled, and the two cancelling one-another out.

The user-setting is stored in zc.cfg as 'epilepsy_flash_reduction'.