Feat: Finish Adding New ZScript Debug Console
Changelog: Added a new ZScript Debug Console. This is attached to another
process, so closing it will not terminate ZC, nor will it terminate the parent process.

It is based on CConsoleLogger / CConsoleLoggerEx.

Traces will be in white. Script IDs green, errors in red.

You can open or close the debugger from the menus, and a config int allows opening
it with the ZC process. You can also manually close the ZScript Debugger without fear.