Feat: Add sprite->Gravity & instrs for mapdata/Screen->SideWarpID[]
Changelog: Added byte obeys_gravity to sprites, and cause weapons to use this.
Old hardcoded falling now instead sets this to 1.
Added ZScript SETTER and GETTER vars as bool T->Gravity to lweapon, eweapon,
npc, and item. The latter two have no effect at this time.

Added int SideWarpID[int direction] to Screen-> and mapdata-> pointers.
The interpreter only has an empty block for Venrob to complete, but the supporting
code for the parser and the interpreter is in place.

(You can issue the instruction, but it always returns 0, and writing to it has
no effect.)

Rob, you wanted to figure out that madness with bits using multiple screen vars,
so, have at you!