Registration: RegistrationVisitor.h Setup
Add files `RegistrationVisitor`(.h and .cpp)
Lay out the groundwork for the new system in RegistrationVisitor.h
Add registration members to class AST
Move some members from SemanticAnalyzer to RecursiveVisitor, for inheritance purposes
Set up `caseRoot`, for handling the main looping through the script. This is unique, and not an override of an existing function like most other cases.