Registration: Fix HEADER_GUARD and Using issues
Using was finding duplicates due to rootscope duplication, when using `always using`. Fixed with `!isRoot()` check.

HEADER_GUARD was checking the same node multiple times, thus disabling entire imports. I don't know how this was causing as few issues as it was.
Fixed this by adding `bool checked` to `ASTImportDecl`. If a decl passes check once, it is marked checked, and will not be checked on subsequent checks.