Parser: Handle global variable initializers
Added ASTExprVarInitializer, as a wrapper node.
ASTDataDecl->setInitializer() still takes ASTExpr, but will now construct an ASTExprVarInitializer from it.
Datum can now give a compiletimevalue for non-constants, but takes a bool for if it should or not- defaults false.
New error, thrown when trying to initailize a global variable to an array.
New error code letter: `R`, for Registration errors.
`Scope* scope` moved from RecusiveVisitor to ASTVisitor, so LValBOHelper can inherit it (which is required for things not to crash)

Now `int a = b; int b = 5;` will work again. In any order.
`int a = 5; const int b = a;` also works, since the initializer is considered compile-time constant.