GlobalSymbols: Fix mapdata pointer
THIS SHOULD NEVER POP TO `NUL`. EVER. WHY IS IT POPPING TO NUL?
THE POINTER SHOULD ALWAYS BE POPPED TO THE REFVAR.
...Uhg.
If this is not popped to `refVar`, then `ri->mapsref` will never be updated, and will not point to the correct mapdata.
This pop is what updates `ri->*ref`, so, popping to `NUL` means not updating the ref. This should only be done in the case of global pointers. But, `refVar==NUL` for global pointers, so popping to `refVar` should always be safe.