Feat: Add `bitmap->isAllocated()`
This checks if a bitmap has been allocated. It will always return true for system bitmap values.
This differs from `isValid()`, in that it will return true even if the bitmap is uninitialized; i.e. cannot be drawn to.
A bitmap which is not valid but has been allocated can be used with `bitmap->Create()` or `bitmap->Read()`
A bitmap which has not been allocated should not be used, and should be reassigned with `Game->AllocateBitmap()` first.