Fix: `Always Deallocate Arrays` QR works better
Now, NEW script types will deallocate regardless of the QR. Also, the deallocation on quest exit will occur regardless of QR.
Global, Item, and FFC scripts will check the QR, and if it is off, behave as in 2.50.2.
There is absolutely no reason that a new script type script should have issues with this, but, prior to this change, they could cause MAJOR array leakage if this was unchecked, which should not be allowed.