Feat: Store ZScript Version on compile.
Changelog: The internal ID of the ZScript version is stored when compiling scripts,
during compilation. The value is stored in the QMisc struct.

Versions prior to 13 read as -1.

This feature should be used with GREAT CARE as individual scripts do not
have this metadata and obviously, there is no way to determine in
what build an ASM script was compiled.

Build is now 47. Updated build date.