Fix: Older ZScript Versions Displayed
Changelog: Versions of Zscript prior to 13 are now shown, rather than -1.
NOTE however, that this merely reflects the constant V_FFSCRIPT used
by ZQ when that quest was saved, and it does not reflect when it was truly
last compiled (13+ do), but it is better than a generic '-1' and
may prove useful.

Linked all checks for this through 'FFCore.quest_format[vLastCompile]'.