Fix: Delayed F6
FFCore.runF6Engine is now called from advanceframe(), unless the third parameter is passed as 'false'.
This should be false for Link OnWin/OnDeath, as well as the F6 script itself.
This is ALSO false for LinkClass::scrollscr(), as that needs custom F6 handling to draw the scrolling screen properly. This is handled via 'FFCore.runF6EngineScrolling()'.
This still needs testing.