Feat: Add 'Game->SaveAndQuit()' and 'Game->SaveAndContinue()'
These do as their names say. They benefit over doing 'Game->Save(); Game->End();' (or 'Game->Save(); Game->Continue();' respectively) in that they will save the game AFTER the execution of the global script onExit.
If an onExit script is important to the quest, these may be neccesary over the prior options for scripts to function as expected.