Attempt fix: Large enemy solidity checks
Check every 8 pixels, with an extra check for the edge of the hitbox (for non-multiple-of-16 hitboxes)
Do various maths to make sure it's using the correct coordinates
Add param `dir` to `m_walkflag`, mandatory. Pass `enemy::dir` to it from everywhere but `canmove`. `canmove` passes `ndir`.
Praying that nothing else got broken by this. -V