Fix: ASkipY Frames Animation
The formula for calculating when to reset a combo animation to the start did not properly account for ASkipY, causing it to use the wrong number of frames.
This has been fixed, though 'qr_BROKEN_ASKIP_Y_FRAMES' being enabled will cause it to use the old behavior.
This QR is NOT yet in the GUI, and needs to be added to the second tab of 'Quest->Rules->Backwards Compatibility', when that gets it's tabpanel.