Feat: Added User Themes
Changelog: Added the ability for the user to define custom GUI themes for
ZC Player, and for ZQuest. The GUI set for the user theme is '99'.

The values for the theme are set in zc.cfg under [Theme] for ZC Player, and
in zquest.cfg under [Theme] for ZQ Editor.