Feat: Add User Themes and Fix Display Grid in Mapscreen Selector
Changelog:
Display a grid in the mapscreen selector box in ZQ if the current
map is an overworld palette rank (0x00, 0x10, 0x20 ...).

Added the ability for the user to define custom GUI themes for
ZC Player, and for ZQuest. The GUI set for the user theme is '99'.

The values for the theme are set in zc.cfg under [Theme] for ZC Player, and
in zquest.cfg under [Theme] for ZQ Editor.