Feat: Add Landmine Powers to Step->Effects Combos
Changelog: Step->Effects combos now have a 'Landmine' flag.
If enabled, when the player steps on one, it will generate either an
eweapon, or an lweapon.
The user can set the weapon type in Attribytes[1]: 0 or invalid default to an eweapon bomb
	explosion.
The user can set the initial direction in Attribytes[2]. It defaults to up.
The user can set the weapon damage in Attributes[0]. The defaut amount is 4.
This combo type will always advance by one (as Next) if activated f the Landmine
	flag is enabled, unless the user also toggles the flag 'Don't Advance'.
	If that flag is enabled, the combo will spawn another weapon when the player
	walks on it.
Finally if the weapon type is set to a wScript (custom) weapon type, the weapon
	that the engine generates will default as an eweapon, unless the user
	toggles the flag 'wCustom is LWeapon'.