Fixed issues with lock blocks;
"Thief" and "No Drain" were off by one; No Drain was using Thief's flag, and Thief was using... whatever the thing right before Thief in the list is I'm tired okay.

"Use Counter" was broken for a myriad of reasons, such as improper if/else positioning, and no default return statement (it was in an "else", but Use Counter went into it's own block and didn't automatically return uggggh).