Added npc->Data.
Added npc->Data as a script variable and as an editor box. This is a ghost exclusive variable intended for usage with ghost. It does nothing besides store a value; I don't see this being a problem. I know this probably goes against your standards Zoria, but my motivation for Ghost 3 hinges on this and if you let this slide I'll owe you one (as in I'll probably do something on your todo list for you)