Remove excessive calls to `rebuild_trans_table` every frame of zquest.
~37% of cpu time when idling zquest was from two calls of this function
per call to `refresh`. It was likely applied to fix a very specific
issue with transparency not updating, but the real fix should instead
be to update the transparency when needed on the appropriate user action

We simply await for the proper user report to know exactly what to fix.

https://discord.com/channels/876899628556091432/995477154164908184