improve scrollscr_butgood / make more accurate
copy/paste:

it was backwards
before, it would load the new screens into memory and draw everything relative to the starting screen, then when scrolling finished it would delete everything and "really" load the screen/region (pretty much, redoing a bunch of loading code). Also, that meant that if scripts written to run during scrolling actually modified the screen contents (which they can) it would get wiped
now it flips it–it stashes all the current screens in a temporary vector, loads the new screens right away (well, after the short "waiting" period), does the scrolling deal (which I now made much smarter to be able to choose between rendering current screens VS the old screens I held on to).... then when the scrolling is done it doesn't need to reload anything
it was really fucking annoying to get it right again. somehow it was different flipping it around, ran into a couple new problems. but also found a couple bugs in the old way
anyway, prob much better now. def. better commented