Reduce excessive codegen in init_data.cpp
For a debug build on my mac, compiling this unit took 30+s,
which was ~60% of the total compilation time for the zelda
target. The culprit was a lot of calls to unique instantiations of
templated functions, one for pretty much every variable this form modifies.

I replaced this with some runtime method calls that accept pointers to the
member variables to modify, and creates the widgets from a handful of common
functions.

I believe `WORD_FIELD` contributed the most to the savings here.

Time to build init_data.cpp.o:

Debug   32s ->  7s            -78%
Release 52s -> 11s            -79%

Size of init_data.cpp.o:

Debug   46364136 -> 10677664  -77%
Release  1678744 ->   581128  -65%