Clear current_subscreen_active on init_game
This fixes a bug where the saved B item (and probably others)
would not be restored from what was saved in the game file. It would
only restore as expected on a fresh restart, when
current_subscreen_active is null, but not when going back to the title
screen from a game and selecting the game file.