win: hide mouse cursor is hw_cursor is disabled
This was being called before but it seems some fault
in the mac driver made it seem `al_hide_mouse_cursor` shouldn't
be called. But in windows the hw cursor shows on top of the allegro
cursor, so seems it's needed.

Need to work out why Mac differs and file a bug with allegro.