save replay even on quit
Previously replays only saved on quit for test mode, but in order
to better handle features like save points - where the game is saved but
doesn't exit - cannot tie replay saving to quitting. Otherwise, the time
between saving and doing an actual "quit game" would not be captured in
the replay; and it wouldn't be 100% right to always exit (and force
going through init_game) when a save point is used during replay.