Fix overflowing version fields
version and minver are both written and stored as fixed-sized strings,
unlike the other strings in the header which are expected to be null
terminated. But code was treating them like null-terminated strings,
causing problems (either borked Quest Info showing the version, or
if built with GCC crashing with a buffer overflow error).

Added better handling for those fields; and also patched up some
similar security vulnerabilities re: the author and title fields.