stop loading quest template to reset items
`reset_items` it called on start up, and it loads the quest template
file to initialize the items to whatever is in there. Except there is
no item section in the classic module's template file, and the module
system is dead weight at this point, so let's stop loading this thing
for no reason.

Benchmarks:
  hyperfine './zelda -test modules/classic/classic_1st.qst 0 119'

Baseline (classic_1st.qst):
  Time (mean ± σ):      4.239 s ±  0.145 s    [User: 0.926 s, System: 0.410 s]
  Range (min … max):    4.133 s …  4.638 s    10 runs

Don't load quest template (classic_1st.qst):
  Time (mean ± σ):      3.852 s ±  0.094 s    [User: 0.710 s, System: 0.403 s]
  Range (min … max):    3.776 s …  4.107 s    10 runs