fix(zq): do not show guys in screen enemy selector
The first twenty enemy ids are reserved for guys, and can produce
unexpected behavior when used as screen enemies. If a screen is using
one of these enemy ids, the name was showing as "(None)" in the screen
enemy list. It will now show "Guy - Do Not Use!".

The enemy selector list also will no longer show any of these enemy ids.