refactor: actually skip reading qst sections when requested
Only the tile importer for qst files uses the feature of skipping
sections when loading a qst file. This would still read and process
every section, it just wouldn't persist anything. So it wasn't really
any quicker.

This paves the way for removing `keepdata` from all of our loading code.