refactor(zc): split save code from title.cpp, remove .icn
.icn files contained just the color data for each ring variant needed
by the title screen. The idea was to avoid loading each quest just to
show a single tile graphic. But, this same data is already in the gamedata
save structure, so we don't need to maintained a separate file doing the same.
The greatly simplifies much of the code dealing with displaying these icons.