feat(zc): split saves into individual files
Each save slot created is now stored in a separate file, kept in the
`saves/` folder. This path is configurable with `[zeldadx] save_folder`.

An existing `zc.sav` file is split into individual files, and the
original is moved to `saves/backup`.

The order of the save slots is maintained by `saves/order.txt`. You may
freely rearrange the lines of this text file, though you must do so when
the program is not open for it to have any effect.

Any save file that gets deleted will simply be moved to `saves/deleted`.
If you ever want it truly gone, you can delete the folder yourself. If
you want to restore it, just move it back to `saves/`.

Individual save files are only loaded into memory as needed, and only
partially if only needed to display info on the title screen (like the
name, icon, hearts, etc).

The `-standalone` switch still works the same, but the save file is at
`saves/standalone.sav`.