feat: add software updater
ZUpdater looks for the newest versions, and allows you to download
and replace your current installation with it.

You can run `zupdater.exe` directly, or you can use the new "Update" tab
in ZLauncher. ZLauncher also has an option to auto-focus the tab on
startup if a new version is detected.

Note: this feature currently requires Python 3 to be installed.

Also: drive-by fixes for launching a process on unix, though the updater
feature doesn't use this.