feat(web): support music crossfading, set position, set speed
This drops the web-specific implementation of zcmusic via SDL_mixer. It
was only done that way because we weren't building any sound libraries
from source, so using emscripten's port system + SDL_mixer was a
workaround. But SDL_mixer only supports a single music channel, so it
cannot do crossfading, so it's time to drop it.