feat(zc): remove encoding layer for save files
Just like the previous work to remove the encoding layer for qst files.

Previous save files can still be read, but when next saved will no
longer be encoded.

Should be a tad faster to read and write, especially reading in bulk as
the title screen no longer needs to read the entire file to decode, when
it only needs the first part of the file.

Also changes save file writing to disk to instead write to a temporary
and move to the destination when done. This may prevent a source of save
file corruption.