refactor: simplify render tree drawing
This adds `prepare` and `render` functions to the RenderTreeItem
class. Also makes them virtual so that new classes can be made to
represent a render item. The minimap drawing code was moved to this.

Consolidates the logic for when a render item is dirty, and moves
allegro 4 bitmap handling into a new derived class.